Apply normalizeQuery to filters with QueryParser
